Replaced unsigned ints in various places with NSUIntegers
git-svn-id: http://basetenframework.org/svn/trunk@1394 fe1aa653-7022-0410-af67-bafd91c7e665

+++ Sources/BXAbstractDescription.h
@@ -31,7 +31,7 @@
 
 @interface BXAbstractDescription : NSObject <NSCoding>
 {
-    unsigned int    mHash;
+    NSUInteger      mHash;
     NSString*       mName;
 }
